+++ CHANGELOG.md
@@ -14,12 +14,16 @@
 * digest
 * restrict comments to themhimselves
 * test deployment to eq.local
+* pyproject.toml refuses version number `*.*.*-SNAPSHOT`
+
 
 
 # 0.3.8-SNAPSHOT
 
+- added `digest.clj`
 - {:class "bg-lime-200 h-40 border-1 p-2 w-2/3"}
-
+- throw exception when not find ruff, python, pytest.
+- throw exception when uploading without choosing a file
 
 # 0.3.7 (2025-09-21)